Merged PR 1052: Recurrent DL Component in TF2X mode
## Implementation Details
To create the recurrent DL component in TF2X mode, I used the Keras layers `simpleRNN()`, `GRU()`, and `LSTM()`. These layers are called depending on what the user passes in from the frontend. The parameters for the activation function are also adjusted to the spec of the Keras layer, where they're all lowercase. This is done by using the string method `lower()`, and it's only executed if the activation function passed in from the frontend is not `LeakyReLU` or `None`. In the event that it's the former, we must pass in the input to another Keras layer because advanced activation functions are separate layers in the Keras API. In the event that it's `None`, we simply set the activation function inside the respective Keras DL Recurrent layer to `None` rather than a string containing that (ie. `'None'`).

After the layer is properly initialized in the `build()` method, it's invoked in the `call()` method. We get the weights from the layer directly using the Keras API, namely the attribute `<Keras Layer Object>.weights`. From there, a list is returned, with the first index being the kernel (mapped to 'W' in the output dictionary), and the last index being the biases.

Depending on whether or not the user elected to dropout, that Keras layer may also be invoked subsequently. In Keras, dropout rate is passed in not as the keep probability, but the actual dropout rate. However, because we pass in the keep probability, I simply changed the dropout rate to `1 - keep_probability`, followed by the multiplier for whether or not the layer is in training mode.

## Where to Start Reviewing
You have to review this branch in TF2X mode, so calling `export TF2X=123` before starting up the kernel is required.

* `/backend/perceptilabs/layers/deeplearningrecurrent/tf2x_template.j2`: Where the layer is implemented.
* `/backend/perceptilabs/layers/deeplearningrecurrent/tf2x_imports.json`: Imports for this layer.
* `/backend/perceptilabs/layers/deeplearningrecurrent/test_deeplearningrecurrent_integration.py`: Test cases were added, however they are simply the original test cases appropriated for TF2x.

Related work items: #1479