Merged PR 1060: Pre-Trained Model: VGG16
Here's what the pre-trained components look like in action in the tool (using the DL component as the temporary design):

![Screen Shot 2021-02-03 at 6.30.04 PM.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1060/attachments/Screen%20Shot%202021-02-03%20at%206.30.04%20PM.png)

## Implementation Details
To implement the VGG 16 pre-trained model, I leveraged the layer spec structure we currently have for every other layer, from which the attributes are passed into the `Keras.applications` API where the pre-trained model is constructed. The `Keras.applications` API, namely the `VGG16()` object is created in the Jinja template, and is callable like every other layer in the tool.

The parameters we include in the spec file are:
* `include_top`: A boolean which decides whether or not to include 3 Dense layers at the beginning of the VGG16 model.
* `weights`: A string which contains the name of the pre-tained weights being loaded into the model. Defaults to `'imagenet'`.
* `input_shape`: Must be at least [64,64,3] for the model to produce output without crashing.
* `classes`: The number of classes, since this is an image classification model.

I also wrote two test cases in the integration test file. One tests the instantiation of the pre-trained model as a layer, the other verifies that the output has the correct shape and that the layer can be run.

## Where to Start Reviewing
* `/backend/perceptilabs/layers/pretrainedvgg16/spec.py`: Review this to see which parameters the user can manipulate in the frontend when that's
* `/backend/perceptilabs/layers/pretrainedvgg16/test_pretrainedvgg16_integration.py`: Where the test cases are written.
* `/backend/perceptilabs/layers/pretrainedvgg16/tf2x_template.j2`: Where the Keras API is invoked and the pre-trained model is instantiated.
* `/backend/perceptilabs/layers/pretrainedvgg16/tf2x_imports.json`: Imports for this layer.

## Other Notes
* `frontend/src/config/dev.env.js`>`ENABLE_TF2X` should be set to "'true'"
* The kernel must be run in TF2X mode, so `PL_TF2X=123 python main.py` to run the kernel

Special thanks to @<David Han> for his help with the frontend work for this story

Related work items: #1565