Merged PR 999: [1405] Fixes access token not being updated in the store
This PR fixes the problem of the access tokens becoming stale and never being refreshed in the store.

One of the potential problems is that if the user questionnaire is active and not completed it before the access token expires, the request to KeyCloak will not work.

The branch this PR references implementes 3 changes:
- updates the the tokens when a refresh is sucessful,
- updates the logic for how offline users can still use the tool, and
- adds error checking for the user profile (secondary to this PR).

**To test**, please:
1. Startup the tool,
2. Register a new user (so we get the questionnaire page),
3. Wait for the token to be invalid (5 minutes in the dev environment),
4. Click on Cancel or try to fill in and Confirm, and
5. Refresh the page

Note: please confirm that  that the offline functionality also works.

---

Before submitting your PR, please make sure that the description contains:
- [x] Implementation details
- [x] Why the change was needed
- [ ] (If applicable) Where to start reviewing and if something needs extra testing
- [ ] (If applicable) Performance changes

And that you:
- [ ] Include any new files in included_files.txt
- [ ] Make sure you can build the wheel and run it
- [ ] Update any relevant Readme's

After the PR has gone through, please:
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #1405