Merged PR 1085: Trainer: Stopping Training
## Implementation Details
---
* `/backend/perceptilabs/trainer/base.py`:
    * `run_stepwise()`: In this method, where the training loop is located, there's a check for if the trainer has been stopped or not.
    * `stop()`: A message indicating a state change to `'Finished'` is logged using the App Logger and the the state variable is changed
* `/backend/perceptilabs/trainer/test_trainer.py`: Unit test for pausing and unpausing is here.

## Why the Change Was Needed
---
The trainer functioned as far as training goes. However other core functionality of the tool, including pausing and stopping training, was not present in the new module. This story aims to add functionality for stopping.

## Where to Start Reviewing and If Something Needs Extra Testing
---
The files listed in the implementation details are where the reviewing should be done.

### Other Notes
To test this PR, the kernel must be run in TF2X mode and the trainer must be activated. This script should suffice: `PL_TF2X=123 python main.py --trainer standard`. It's also advised that the data wizard is active, so `frontend/src/dev.env.js` > `ENABLE_TF2X` and > `ENABLE_DATA_WIZARD` should both be true.

Related work items: #1544