Merged PR 1086: Trainer: Pausing Training
## Implementation Details
---
* `/backend/perceptilabs/trainer/base.py`:
    * `run_stepwise()`: In this method, where the training loop is located, the method for performing the actual pausing is called.
    * `_sleep_while_paused()`: This method sleeps for one second using the `time` library (as we've done before for pausing) as long as the state of the trainer is paused.
    * `pause()`: A message indicating a state change to `'Paused'` is logged using the App Logger and the the state variable is changed
   * `unpause()`: A message indicating the state changed from `Paused` to whatever the previous state was (which is stored in a new member variable) is logged using the App Logger.
* `/backend/perceptilabs/trainer/test_trainer.py`: Unit test for pausing and unpausing is here.

## Why the Change Was Needed
---
The trainer functioned as far as training goes. However other core functionality of the tool, including pausing and stopping training, was not present in the new module. This story aims to add functionality for pausing.

## Where to Start Reviewing and If Something Needs Extra Testing
---
The files listed in the implementation details are where the reviewing should be done. There is a slight quirk with the pausing functionality in the frontend, the kernel changes the state of the trainer correctly although there's a delay in the frontend (the pause button is delayed turning into a play button when the kernel is paused). Pressing the pause button again before it changes to a play button results in the kernel to hang. However, if the pause button is pressed once and the delay is observed, then the pause functionality should function correctly.

### Other Notes
To test this PR, the kernel must be run in TF2X mode and the trainer must be activated. This script should suffice: `PL_TF2X=123 python main.py --trainer standard`. It's also advised that the data wizard is active, so `frontend/src/dev.env.js` > `ENABLE_TF2X` and > `ENABLE_DATA_WIZARD` should both be true.

Related work items: #1543