Merged PR 1092: Trainer: Computing Total Duration
## Implementation Details
---
* `/backend/perceptilabs/trainer/base.py`: The training duration is computed here. This is done using the time library at the start of each epoch and at the end of each epoch, similarly to how it was done in the previous training module with training components.
    * `_sleep_while_paused()`: In this method, the time paused is computed and returned. It's subsequently subtracted from total epoch time, and consequentially, from total training time.

## Why the Change was Needed
---
Measuring training time is integral in gauging the performance of ML models. As a modeling tool, this was a simple but important feature for us to implement.

## Where to Start Reviewing
---
The files listed in the implementation details ought to be the place to begin reviewing. To test this PR in the tool, run the kernel using the following script: `PL_TF2X=123 python main.py --trainer standard`. It's also advised since the trainer works well in conjunction with the Data Wizard to set the appropriate `dev.env.js` variables to true, similar to all other trainer related PRs thus far.

### Other Notes
This PR utilizes pausing in the trainer, which is implemented in [this](https://dev.azure.com/PerceptiLabs/PerceptiLabs/_git/PerceptiLabs/pullrequest/1086) PR.

Related work items: #1569