Merged PR 1095: Add validation phase to trainer
# Implementation Details
---

Most of the work was to refactor the data loader so that it supports 3 different splits of data. Likewise, we now create 3 sets of processing pipelines: training (preprocessing), inference (preprocessing for validation and test) as well as one general set for postprocessing (it may be split later if we should need it).

Finally, the trainer now does a pass over the validation set for each epoch.

Since this is a prerequisite for other stories (see below) and we don't have a design for now, the values are unfortunately hardcoded. Frontend work will be done in: https://dev.azure.com/PerceptiLabs/PerceptiLabs/_workitems/edit/1469/?workitem=1667 once a design is ready.

# Why the change was needed
---

This PR implements the architectural changes needed for splitting the dataset and running a validation phase in the trainer. Therefore, it is necessary for completing the following stories:

1540 - As a user, I want to see the accuracy during training
1571 - As a user, I want to see the loss in the frontend

# How to test

When training, you should see that the progress bar switches between training and validation.

To run the tool:

* In frontend/src/config, set ENABLE_TF2X and optionally ENABLE_DATA_WIZARD to True
* Run the Kernel like so: PL_TF2X=123 python main.py --trainer standard

# Please review the following files

**/backend/perceptilabs/data/pipelines/__init__.py, /backend/perceptilabs/data/pipelines/image.py, /backend/perceptilabs/data/pipelines/numerical.py**
* Return the 3 different flavors of pipelines
* Rename from preprocessing -> pipelines

**/backend/perceptilabs/data/base.py**
* Significant refactor before adding the split functionality

**/backend/perceptilabs/data/test_data_loader.py**
* Adds tests for splitting

**/backend/perceptilabs/trainer/base.py**
* Adds validation phase to training loop

**/backend/perceptilabs/trainer/test_trainer.py**

* Modifies tests to account for validation phase

** /backend/included_files.txt **

* Change to new file names

Related work items: #1537