Merged PR 1109: check if vars are trainable
# Implementation Details
-------------------------------
## What was changed?
The TF2X base layers will no longer perform weight updates on non-trainable weights

## Why was it changed?
To ensure that we don't change the gradients of pre-trained layers.

## How was it changed?
Checking if the trainable properties is set to false. If so, the gradient will be skipped by the training layer when it updates the gradients.

# How to review
-------------------------------
Create a network like this and load some image data:

![image.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1109/attachments/image.png)

# Please review the following files
-------------------------------
/backend/perceptilabs/core_new/layers/base.py

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Added docstrings to the properties

#Please also:
- [X] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

check if vars are trainable