Merged PR 1111: Story 1538: Export Functionality for Trainer
# Implementation Details
-------------------------------
## What was changed/Why was it changed?
The main purpose of this was to add export functionality for the trainer. This was done by creating the exporter module, located in `/backend/perceptilabs` next to the trainer. It's designed to run independently from the trainer as well as in conjunction with it, hopefully to make exporting a simple operation as a whole when we transition to TF 2X fully.

Unfortunately, loading the inference model where the outputs and inputs are pre-defined doesn't work for the way we wanted to load checkpointed models from disk. Because of this, we use `save_weights()` rather than simply calling `save()`.

## How was it changed?
It was changed by changing the logic for how exports are handled from the `mainInterface` down to the exporter module. In TF2X, we no longer create coreLogic for when the only operation is export, rather, the exporter is called directly.

# How to review
-------------------------------
Run the kernel with `PL_TF2X=123 python main.py --trainer standard` and enable both the Data Wizard and TF2X in `dev.env.js`.

To test export functionality:
1. Start training a model
2. Try exporting to a directory
3. Let it finish training
4. Check `documents/perceptilabs/default/<model_name>/checkpoint`, there should be checkpoint related files related to saving weights.

# Please review the following files
-------------------------------
**Top of the execution flow:**
* `/backend/perceptilabs/mainInterface.py`: Changes to support both tf1x and tf2x export requests, done through code refactors. The trainer is invoked at this level if export is called **after** training
* `/backend/perceptilabs/coreInterface.py`: Forward export request sent from mainInterface to  the trainer, only reach this point if an export request is sent during training.

**Exporter:**
* `/backend/perceptilabs/exporter/base.py`: The exporter module is defined here. We load the checkpoint model from disk here so it can be exported by the user using a method that's also defined here. Also include functionality for getting the inference model, exporting the inference model should be trivial for a future story as a result of this.
* `/backend/perceptilabs/exporter/test_exporter.py`: Unit tests for exporter.

**Trainer:**
* `/backend/perceptilabs/trainer/base.py`: Invoke the exporter at the end of training.
* `/backend/perceptilabs/trainer/test_trainer.py`: Integration tests for the exporter and the trainer, one to see if exporting in the middle of the training loop works, and another to see if exporting after training works.

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
* Extracted the code for create_export_response into a method, improving readability
* Created the exporter, added unit tests for durability
* Removed commented code

Related work items: #1538