Merged PR 1124: Story 1711: Dice Loss Function for Trainer
# Implementation Details
-------------------------------
## What was changed?
An extra case was added to the conditional where the loss function is resolved. This is to add functionality for the Dice loss function in the kernel. There's also a new python file in the trainer module where the dice function is defined (as well as the weighted cross entropy loss function, but that's not included in this PR for reasons involving the trainer). Test cases were also added to see if the custom losses work properly.

## Why was it changed?
It was changed because Dice loss is a commonly used loss function, and in order to restore as much of the previous functionality of the old workflow in the new trainer, it was added to the trainer module.

## How was it changed?
Adding Dice was simple– it was already written for the tf2x classification layer (old workflow), so adding it to the trainer was as simple as copy and pasting. To ensure that it's actually invoked, adding a conditional for when its part of the training settings is added in the frontend.

# How to review
-------------------------------
This PR uses custom settings, which is part of [this](https://dev.azure.com/PerceptiLabs/PerceptiLabs/_git/PerceptiLabs/pullrequest/1120) PR.

1. Start the kernel using `PL_TF2X=123 python main.py --trainer standard`
2. Change the following environment variables in `dev.env.js` to true:
    * `ENABLE_TF2X`
    * `ENABLE_DATA_WIZARD`
    * `ENABLE_GLOBAL_TRAINING_SETTINGS`
3. Load [data.csv](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1120/attachments/data.csv)
4. Press `Run` in the tool with the `Dice` loss selected (**NOTE**: Be sure to set the batch size to a small number like 1 or 2. Otherwise the visualizations don't work in the frontend)

# Please review the following files
-------------------------------
* `/backend/perceptilabs/trainer/base.py`: This is where the dictionary is resolved from strings into real methods from the Keras library, and in the case of this PR, into the custom loss function defined in `losses.py`.
* `/backend/perceptilabs/trainer/test_trainer.py`: Updated unit tests with training settings fixture.
* `/backend/perceptilabs/trainer/losses.py`: Dice is defined here.

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
* Extracted loss functions into a separate file to maintain the culture of readability (since it's easily imported in the trainer and returned when it's resolved to that loss) and durability (since it's in a separate file, adding to it and making changes should be easy) in the codebase.

Related work items: #1711