Merged PR 1099: [522] Toggle between features using a slider
## Why the Change Was Needed
___
Currently, only one feature map can been seen for a convolution layer.
This PR adds the ability to view the other feature maps, with the help of a slider.

## Implementation Details
___
The kernel now returns all of the features maps for the frontend chart requests (stored within the layer itself)

We now have a `chartIdx` property in the layerMeta object, so that the previews in the workspace and in the sidebar can be synched.

We also a new getter/setter in the `mod_events` module so that the slider button/thumb can be used without dragging the layer itself.

## Where to Start Reviewing and If Something Needs Extra Testing
___
Please make sure that:
- the arrows in the chart previews work,
- the slider works,
- the tooltip displays the correct value, and
- the drag/drop function of the layer still works.

This is what the feature looks like for me:
![522_featureMapSlider.gif](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1099/attachments/522_featureMapSlider.gif)

## Notes
___
Kudos to @<Adil Salhotra>  for doing all the hard bits :pray:

Related work items: #522