Merged PR 1132: Training Settings bug fixed when running in TF1X mode
# Implementation Details
-------------------------------
## What was changed?
A small change was made to the mainInterface because training settings don't exist in the same fashion in TF1X mode as they do in the new trainer. This was causing a keyError in the mainInterface.

## Why was it changed?
It was changed so the kernel could start training in TF1X mode.

## How was it changed?
A simple check is provided to see if the key 'trainSettings' exists, if it does, then assign it to it's correct value counterpart by using `value['trainSettings']`, otherwise, pass in `None` for the training settings.

# How to review
-------------------------------
Try running the kernel in regular TF1X Mode and start training. Also, start training the kernel in TF2X mode with all of the Data Wizard, TF2X, and Training Settings environment variables set to `True`.

# Please review the following files
-------------------------------
`backend/perceptilabs/mainInterface.py`: Where the changes detailed above were made.

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Hopefully this made the TF2X implementation more durable since running TF1X mode is no longer affected by it. Avoiding any negative trickle down effects from adding to the codebase is important.