Merged PR 1147: Extra long content overflowing layer preview container
# Implementation Details
-------------------------------
## What was changed?
The sizes of the containers around the chart-picture component.

## Why was it changed?
When dealing with data that was smaller than the container, it would be scaled up to take up all the space available.

The previews didn't need to work with data having dimensions larger than the container, hence the container's permission for the container to be given all the space it needed (until the width is maxed out).

The previews looks as follows:
![image.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1147/attachments/image.png)

## How was it changed?
Heights were set.

It currently looks as such in the modeling workspace:
![image (2).png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1147/attachments/image%20%282%29.png)

And in the statistics view:
![image (3).png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1147/attachments/image%20%283%29.png)

Checking that previous fix still works:
![image (4).png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1147/attachments/image%20%284%29.png)

# How to review
-------------------------------
Please:
- enable
`ENABLE_TF2X`, `ENABLE_DATA_WIZARD`, and `ENABLE_GLOBAL_TRAINING_SETTINGS` in `dev.env.js`
- run `PL_TF2X=123 python main.py --trainer=standard` to start the kernel,
- create new model,
- load Robert's `Documents.zip` file (found [here](https://www.notion.so/Shared-files-001c64dd2c444cf6818a69be6c7a3205#26d24f591e3b4566ab7356c71e6fb92e)),
- set input column to `Image` and the output to `Category`