Merged PR 1131: Layer preview not always shown when reloading the app
# Implementation Details
-------------------------------
## What was changed?

The check to see if the preview has been fetched is removed.

## Why was it changed?

Due to an issue related to a race condition between loading the model from disk and loading from IndexedDb, the modeling tool may think that a preview has already been fetched when it hasn't been.

# How to review
-------------------------------

With a number of model open, please:
- start a training session,
- reload the tool once the training has started.

The preview should be fetched for every tab change.

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------

- fixed spelling error
- added error checking