Merged PR 1173: Set the right link for documentation page.
# Implementation Details
-------------------------------
## What was changed?
It was changed redirect url to ducumentation page.

## Why was it changed?
To not redirect to 404 page

## How was it changed?
Right url was set.

# How to review
-------------------------------
Click on top-right questionmark then on Documentation link

# Please review the following files
-------------------------------
*To be filled in*

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
I extracted the url in constants and replace it in all places to have same url.

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Set the right link for documentation page.

Related work items: #1811