Merged PR 1178: Allow creating model when tf feature flag is enabled
# Implementation Details
-------------------------------
## What was changed?
It was made enable create button when only tf2 env variable is enabled

## Why was it changed?
Because it can't create model beacuse button was dissabled.

## How was it changed?
It was verified if data wizard env var is enabled and base on that decide if make verification or not.

# How to review
-------------------------------
Enabled tf2 and dissabled data wizard and data setting env variable.

Allow creating model when tf feature flag is enabled

Related work items: #1817