Merged PR 1176: PerceptiLabsVisualizer used by all Inner Layers
# Implementation Details
-------------------------------
## What was changed? Why was it changed?
The PerceptiLabs visualizer is added to the Deep Learning components and the Pre-Trained components in the tool. Doing this is the baseline requirement for integrating the PL Visualizer in the tool.

## How was it changed?
To implement the `PerceptiLabsVisualizer` module in all inner layers, it's added to the imports and its method `visualized_trainables` is added to all Deep Learning and Pre-Trained layers.

# How to review
-------------------------------
* Start the kernel using `PL_TF2X=123 python main.py --trainer standard`
* Change the following environment variables in `dev.env.js` to true:
     * `ENABLE_TF2X`
     * `ENABLE_DATA_WIZARD`
     * `ENABLE_GLOBAL_TRAINING_SETTINGS`

Run a model using the following datasets:<  [multichannel_75x75.zip](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1176/attachments/multichannel_75x75.zip),  [multichannel_299x299.zip](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1176/attachments/multichannel_299x299.zip) >. The former dataset is to be used with Resnet50 and VGG16, the latter is meant to be used with InceptionV3 and MobilenetV2.

# Please review the following files
-------------------------------
In `layers` directory in the backend, look at the following files in `deeplearningrecurrent`, `pretrainedinceptionv3`, `pertainedmobilenetv2`, `pretrainedresnet50`, `pretrainedvgg16`:

* `tf2x_template.j2`
* `tf2x_imports.j2`

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
The improvements are durable because they utilize the `PerceptiLabsVisualizer` module. The more we modularize operations that happen within the codebase, the more durable the codebase is as a whole.

**Note:** Using the 299x299x3 image data results in the frontend lagging quite a bit. However the training loop starts and finishes as expected.

Related work items: #1657