Merged PR 1232: Visualizations out of sync
# Implementation Details
-------------------------------
## What was changed?
statistics for iooutput layer

## Why was it changed?
visualizations for iooutput layer and other layers are out of sync.

## How was it changed?
final sample in the batch will be used  for statistics instead of first sample in the batch.

# How to review
-------------------------------
start perceptilabs with data wizard settings set to true.
create any model with categorical outputs.
start training and check and the metrics.
make sure, the ioinput metrics and iooutput metrics are in sync.

# Please review the following files
-------------------------------
all

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
added comments to make  bugs don't get repeated.

#Please also:
- [X] Include any new files in included_files.txt
- [X] Update the story on the board
- [X] Announce changes that will affect everyone in #dev in slack

fixing the bug

Related work items: #1804