Merged PR 1217: Monitor final loss and accuracy/iou
# Implementation Details
-------------------------------
## What was changed?

Added support for metrics in the MixPanel events training-stopped and training-completed.

Each event gets a set of properties like below:

![image.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1217/attachments/image.png)

I.e., we compute the max/min/average over all layers that have a loss value. The same goes for accuracy or IoU

## Why was it changed?

We want to see if a user aborts because the model performed poorly.

The reason we use max/min/avg/num is because (a) several layers can produce the same metric [e.g., two categorical layers will produce accuracy] and (b) MixPanel works best with simple key/value events.

## How was it changed?

* Added a summary to the IoOutput stats objects. This is where final metrics can be declared
* Added tracking for individual losses in the trainer (i.e., per layer losses)

# How to review
-------------------------------
1. start training a network with categorical/image/numerical outputs
2. let training complete or stop it manually
3. monitor the MixPanel live feed

Note: if you stop it manually, the final metrics wont perfectly match those in the frontend. It seems like the frontend stops rendering the data immediately, but it takes a few more seconds before the trainer receives the request. You can compare the progress bar and the progress recorded in the event to verify this.

# Please review the following files
-------------------------------

**/backend/perceptilabs/layers/iooutput/stats/categorical.py**: Added _get_summary_method, removed deprecated method __get_dataobj_loss_
**/backend/perceptilabs/layers/iooutput/stats/image.py**: added _get_summary_method
**/backend/perceptilabs/layers/iooutput/stats/tracker.py**: added Loss tracker
**/backend/perceptilabs/stats/accuracy.py**: added method to get final metric value per phase
**/backend/perceptilabs/stats/loss.py**: added method to get final metric value per phase
**/backend/perceptilabs/stats/iou.py**: added method to get final metric value per phase
**/backend/perceptilabs/tracking/base.py**: added logging while in dev mode
**/backend/perceptilabs/tracking/training_completed.py** Added metrics to payload
**/backend/perceptilabs/tracking/training_stopped.py**: Added metrics to payload
**/backend/perceptilabs/tracking/utils.py**: Added function to compute max/min/avg of a metric that exists in multiple output layers
**/backend/perceptilabs/trainer/base.py**: Added support for tracking per-output losses, removed dead code
**/backend/perceptilabs/utils.py**: Added an is_dev method

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Removed dead code in the trainer and in the categorical stats object. Added test cases for changes to the accuracy/loss trackers

#Please also:
- [X] In...