Merged PR 1219: Set the right content of charts for global view.
# Implementation Details
-------------------------------
## What was changed?
It was changed the content of global loss charts

## Why was it changed?
It was changed because it was displaying wrong data.

## How was it changed?
It was changed data between loss over steps and loss over epochs

# How to review
-------------------------------
Enable data wizard feature flag.
Run model
Select a global tab.

# Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Was set the unique keys to I/O chars.

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Set the right content of charts for global view.

Related work items: #1902