Merged PR 1236: "When using the standard trainer, dataset settings must be set!" mainInterface bug resolved
# Implementation Details
-------------------------------
## What was changed?
To fix this bug, changes in the `mainInterface` and the `mod_api` were made.

## Why was it changed?
The `mainInterface` had a few notable issues, every time an instance of the lightweight core is created it requires the dataset settings to be passed in as well. There were a few responses in the main interface triggered by the `API_getPartitionSummary` and `API_getDataMeta` requests from the frontend that didn't contain the dataset settings. Since the dataset settings are required for the creation of lw core objects, those requests were changed in the frontend.

## How was it changed?
To make these changes, I passed in the dataset settings from the frontend into the `value` dictionary in the kernel. The dataset settings are then passed into the responses (the helper method `_create_response_get_partition_summary`) to the requests that were causing the bug.

# How to review
-------------------------------
To review, use the reproduction steps in the bug ticket.

# Please review the following files
-------------------------------
* `/backend/perceptilabs/mainInterface.py`

* `/frontend/src/renderer/store/modules/mod_api.js`

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
To make the codebase a better place, I introduced a refactor in the mainInterface to help future viewers debug potential issues with the requests `"getDataMeta"` and `"getPartitionSummary"`. I created helper methods at the bottom of the file `mainInterface.py` that are called by the method `_create_response()`.

Related work items: #1814