Merged PR 1235: now generates clean name with their own counters
# Implementation Details
-------------------------------
## What was changed?
Components are now generated with names that are in line with the non-generated components.

## Why was it changed?
Having intuitive names on components are important for first time users.

## How was it changed?
There is now a name counter and a type-to-name-dict which helps us generated components with clean names.

# How to review
-------------------------------
Create any model and verify that the names are along the lines of "Input_1", "Output_1", etc.

# Please review the following files
-------------------------------
backend\perceptilabs\graph\builder.py

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Added slightly to the docstrings in the file.

#Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

now generates clean name with their own counters

Related work items: #1895